pythonic-doom-fire: add ColorPalette class
This is responsible for a nice and flexible way to attributing colors
for an object that uses it.

In future I pretend to implement some linear interpolation here for
adding some fancy color effects, just like smoothly changing from the
current color palette to a target color palette. Some ideas would be
welcome :)