Implementation "doom-fire-algorithm" in Python 3 with PyGame
It was implemented in Python 3, using the PyGame graphical library, the "doom-fire-algorithm". This algorithm was designed in: fire rendering, fire data structure and fire intensity propagation. And it can be fully configured to change the grid size of "pixels", and even the size of the "pixels" themselves.
In the graphical rendering, a technique of creating the "pixels" was used from the drawing of squares, by means of the function "pygame.draw.rect".